Add index reduction when lst->idx wraps around
We were reducing indices (lst->idx and the pivot stack) when extending
the array, but one could have an LST that reaches equilibrium, repeatedly
inserting and popping so that it doesn't have to extend, but which will
keep lst->idx and the pivot indices growing without bound, so they'd
eventually overflow.

To prevent this, we also reduce indices when lst->idx "wraps around", i.e.
increases to a value that reduces to 0 (aka is congruent to zero modulo
the lst's capacity). That's not very often (capacity starts out at 2048
and doubles with each expansion, just like heaps).